Extract tests for models.Requires
Also fix the class and tests.
